+++ package.json
@@ -8,7 +8,6 @@
     "start": "next start",
     "lint": "next lint",
     "postinstall": "prisma generate"
-
   },
   "dependencies": {
     "@prisma/client": "^5.14.0",
@@ -17,13 +16,15 @@
     "jsonwebtoken": "^9.0.2",
     "mongodb": "^6.6.2",
     "morgan": "^1.10.0",
-    "next": "14.2.2",
+    "next": "^14.2.3",
     "next-client-cookies": "^1.1.1",
     "react": "^18",
     "react-cookie": "^7.1.4",
     "react-dom": "^18",
     "react-router-dom": "^6.23.1",
-    "universal-cookie": "^7.1.4"
+    "universal-cookie": "^7.1.4",
+    "webpack": "^5.91.0",
+    "webpack-cli": "^5.1.4"
   },
   "devDependencies": {
     "@types/jsonwebtoken": "^9.0.6",
