+++ src/app/principal/dynamicButton.tsx
@@ -3,11 +3,13 @@ import React, { useState, useEffect, useRef, use } from 'react';
 import MenuDesplegable from './menuDesplegable';
 import MenuAsignar from './menuAsignar';
 import Mensajeria from './menuChats';
+import Recursos from './recursos';
 import { GuardarEdificio, getBuildingsByUserId, builtEdificio, getUEbyUserId, getUEById } from '../../services/userEdificios';
 import { getUserByCooki, getUser, getUserByHash, updateUserBuildings, updateUser} from '@/services/users';
 import {recolectarRecursos, calcularMadera, calcularPiedra, calcularPan } from '@/services/recursos';
 import { getChats, getUsernameOther, getChatName } from '@/services/chats';
 import { getMensajes } from '@/services/mensajes';
+
 
 
 type Building = {
@@ -38,16 +40,6 @@ const DynamicBuildings: React.FC = () => {
   const[menuButton, setMenBut] = useState(false);
   const[idUEClick, setIdUEClick] = useState("");    //id de userEdificios seleccionado ante un click
   
-// VARIABLES PARA LA RECOLECCION DE RECURSOS AUTOMATICA
-const [maderaPorSegundo, setMaderaPorSegundo] = useState(0);
-const [piedraPorSegundo, setPiedraPorSegundo] = useState(0);
-const [panPorSegundo, setPanPorSegundo] = useState(0);
-const maderaRef = useRef(madera);
-const piedraRef = useRef(piedra);
-const panRef = useRef(pan);
-
-
-
 ////-----------------------------------------------------------
 //---------------------seba--------------------------------------
 //-----------------------------------------------------------
@@ -106,63 +98,9 @@ const [message, setMessage] = useState('');
   }, [userId]);
 
 
-  //useffects recursos automaticos
- //useffect para recolectar recursos automaticamente CAMBIAR 50 POR 5
- useEffect(() => {
-  const fetchResource = async (calculateFunc: (id: string) => Promise<number>, setFunc: (value: number) => void) => {
-    try {
-      const result = await calculateFunc(userId);
-      setFunc(result);
-    } catch (error) {
-      console.error(`Error fetching resource: ${error}`);
-    }
-  };
-
-  if (userId) {
-    Promise.all([
-      fetchResource(calcularMadera, setMaderaPorSegundo),
-      fetchResource(calcularPiedra, setPiedraPorSegundo),
-      fetchResource(calcularPan, setPanPorSegundo),
-    ]);
-  }
-}, [userId]);
-
-useEffect(() => {
-  cargarUser();
-  const timer = setInterval(() => {
-    setMadera(madera => madera + maderaPorSegundo);
-    setPiedra(piedra => piedra + piedraPorSegundo);
-    setPan(pan => pan + panPorSegundo);
-  }, 2000);
-
-  return () => clearInterval(timer);
-}, [maderaPorSegundo, piedraPorSegundo, panPorSegundo]);
-
-useEffect(() => {
-  maderaRef.current = madera;
-}, [madera]);
-
-useEffect(() => {
-  piedraRef.current = piedra;
-}, [piedra]);
-
-useEffect(() => {
-  panRef.current = pan;
-}, [pan]);
-
-useEffect(() => {
-  const timer = setInterval(async () => {
-    try {
-      await updateUser(userId, { madera: maderaRef.current, piedra: piedraRef.current, pan: panRef.current });
-      console.log('recursos actualizados');
-    } catch (error) {
-      console.error(`Error updating user: ${error}`);
-    }
-  }, 5000);
-
-  return () => clearInterval(timer);
-}, [userId]);
-
+  //useffects recursos automaticos -> BORRADOS AHORA ESTAN EN EL COMPONENTE RECURSOS
+ 
+ 
 //conseguir todos los nombres de los chats
 useEffect(() => {
   if (chats.length > 0 && userId) {
@@ -477,12 +415,18 @@ const handleBuildClick = async (id: string, x: number, y: number, buildingType: 
   return (
     <div className="hola flex flex-col items-center justify-center w-screen h-screen bg-gray-900">
       <div className="absolute top-0 left-0 p-4 bg-red-500 text-blue font-bold py-2 px-4 rounded">
-      <h3>Usuario: {usuario}</h3>
-        <h3>Madera: {madera} || PS: {maderaPorSegundo}  </h3>
-        <h3>Piedra: {piedra} || PS: {piedraPorSegundo}  </h3>
-        <h3>Pan:    {pan}    || PS: {panPorSegundo}     </h3>
-        <h3>Trabajadores disponibles: {unidadesDisponibles}</h3>
-        <button onClick={() => recolectarRecursosUser()}> Recolectar Recursos</button>        
+      <Recursos 
+        usuario={usuario}
+        userId= {userId}
+        madera={madera}
+        setMadera={setMadera}
+        piedra={piedra}
+        setPiedra={setPiedra}
+        pan={pan}
+        setPan={setPan}
+        unidadesDisponibles={unidadesDisponibles}
+        cargarUser={cargarUser}
+      />         
       </div>
       <div className='absolute top-0 left-100 p-4 bg-red-500 hover:bg-blue-700 text-blue font-bold py-2 px-4 rounded'>
         <button onClick={() => handleMensajeria()}>Chat</button>
